requirements.txt
aiogram==2.14
python-dotenv